signout api admin deshboard

+++ src/actions/auth.actions.js
@@ -59,11 +59,20 @@ export const isUserLoggedIn = () => {
   };
 };
 
-export const Signout = () => {
+export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
-    dispatch({
-      type: authConstants.LOGOUT_REQUEST,
-    });
+    dispatch({ type: authConstants.LOGIN_REQUEST });
+    const res = await axios.post("/admin/signout");
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    } else {
+      dispatch({
+        type: authConstants.LOGIN_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
